iox-#996 Split tests into multiple files to improve readability and reviews
